requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
pytest==8.4.2
pytest-flask==1.3.0
requests==2.32.5
gunicorn==23.0.0
pyyaml==6.0.3